"Refactor retweet controller and add helper function for retweet queries"
This commit refactors the retweet controller by moving the validation logic to the top of the function and removing unnecessary code. It also adds a new helper function `createQueriesRetweet` for creating queries related to retweet functionality, but the function body is still empty and needs to be implemented. The commit does not change any existing functionality related to retweets.