"Remove unused code for deleting retweet in retweet controller"
This commit removes the unused code for deleting a retweet in the retweet controller. The code for finding and deleting a retweet, as well as the related conditional logic and update queries, have been removed. The code for creating a new retweet and updating the user and tweet models remains unchanged.