"Remove ReTweet model, add retweets array to User model and update tweets.js controller"
This diff removes the ReTweet model and its associated schema from the model/ReTweet.js file. It also adds a 'retweets' array to the User schema in the model/User.js file, which will store the tweets that the user has retweeted. Finally, the diff updates the tweets.js controller to reflect these changes by removing all references to the ReTweet model and using the 'retweets' array in the User model instead.