"Refactor tweet creation and retweet functionality in tweets.js controller; add originalTweet field to Tweet schema"
The diff includes changes to the `tweets.js` controller file, refactoring the `create` and `retweet` functions. In the `create` function, the tweet content is now extracted directly from `req.body.tweet`. In the `retweet` function, the tweet is fetched using `Tweet.findById(id)` instead of searching by `postedBy` as well.

Additionally, a new field `originalTweet` has been added to the `Tweet` schema in `Tweet.js`. This field is of type `Schema.ObjectId` and references the `Tweet` schema itself.