"Add function to propagate likes to original tweets in retweet chain"
This commit adds a new function `propagateLikes` to the `tweets.js` controller for the tweet feature. This function is used to propagate likes to the original tweets in the chain when a user retweets a tweet. The function is called after creating a new retweet and recursively goes through the original tweet's chain to add the user to the likes of each original tweet.