"Add userRetweetedTweets field to User model and modify likeTweet controller to consider tweet likes"
This commit includes changes to both the User model and the tweets controller for likes. In the User model, the 'likedTweets' field has been renamed to 'retweetedTweets'. In the tweets controller, the 'likeTweet' function has been updated to check if the user has already liked or unliked the tweet using 'user.likedTweets.includes(id)' and 'tweet.likes.includes(id)'. The 'option' variable is now determined based on the new condition.