"Update tweets.js and helperFunc.js: remove retweetQuery, add isRetweet and retweetId to generateTweetQueries; update TweetQuery in retweet function"
The diff includes changes in two files: tweets.js and helperFunc.js. In tweets.js, the retweetQuery has been removed from the generateTweetQueries function call in the likeTweet function. In the retweet function, the generateTweetQueries function call has been updated to include isRetweet and retweetId as parameters, and the TweetQuery and UserQuery have been swapped in the Promise.all call. In helperFunc.js, the generateTweetQueries function has been updated to accept isRetweet and retweetId as parameters and to conditionally update TweetQuery based on isRetweet.