"Update likeTweet controller to remove unnecessary updatedUser from handleRetweet; update helperFunc to remove unused updatedUser variable in handleRetweet"
The diff includes changes to the controller/tweet/tweets.js and utils/helperFunc.js files. In the likeTweet controller, the updatedUser variable is no longer used after the handleRetweet function call, so it has been removed. In the helperFunc file, the updatedUser variable is no longer assigned a value in the handleRetweet function, so it has been removed as well. Additionally, the updatedTweet variable is now being assigned the result of the findByIdAndUpdate method in the helperFunc file.