"Refactor handleRetweet function and add updateRetweetLikes utility function"
The diff includes changes to the `controller/tweet/tweets.js` file and the `utils/helperFunc.js` file.

In `controller/tweet/tweets.js`, the `handleRetweet` function is called with a new variable `up`.

In `utils/helperFunc.js`, the `handleRetweet` function is updated to include a conditional check for the presence of an `originalTweetId`. If it doesn't exist, the function directly updates the likes on the tweet and returns. If it does exist, the function calls a new utility function `updateRetweetLikes` to update the likes on the retweeted tweets and the original tweet.

The `updateRetweetLikes` function is a new function that takes in the `originalTweet`, `option`, and `userId` as arguments. It iterates over the retweet IDs of the original tweet and updates the likes on the retweeted tweets and the original tweet. If there are still retweeted tweets, the function recursively calls itself to update the likes.

The `updateRetweetLikes` function replaces the previous `updateRetweetLikes` function, which has been removed.