"Convert tweet rendering and addition functions to async, update retweet check in renderTweet"
This commit converts the AddTweet and ShowTweets functions in tweetHandlers.js to async functions. It also updates the renderTweet function to check if a tweet has been retweeted by using the getRetweetInfo utility function. The getTweets and tweetCreation functions in apiOperations.js have been updated to await the ShowTweets and AddTweet functions respectively.