"Update Like.js and tweetHandlers.js: add console log, adjust retweeted check"
This diff includes changes to two files: Like.js and tweetHandlers.js.

In Like.js, a console log statement was added to log the value of `newTweetRetweeted`. In tweetHandlers.js, a console log statement was added to log the value of `tweet`, and the way the `isRetweeted` variable is calculated was changed to check if the `originalTweet` property exists on the tweet object instead of checking if the `isRetweeted` property exists. Additionally, the content of the tweet is updated to include "retweeted" and the username of the retweeter. The retweet count is also updated to be an empty string if there are no retweeters.