"Update Like.js and tweetHandlers.js: added console logs and fixed bug in updatedUiRetweeted function"
This commit updates the files Like.js and tweetHandlers.js by adding console logs for debugging purposes and fixing a bug in the function updatedUiRetweeted. The bug was caused by the incorrect use of the "forEach" method, which should return an array of elements but was instead returning "undefined". The function has been updated to use the "map" method instead, which correctly returns an array of elements.