"Log tweet author in tweetHandlers.js and pass user object to ShowTweets in apiOperations.js"
This diff includes changes to two files: tweetHandlers.js and apiOperations.js. In tweetHandlers.js, the code logs the tweet author in the renderTweet function. In apiOperations.js, the code passes the user object to the ShowTweets function, and assigns the user object to a variable called 'user'. Additionally, the code checks if the 'parentTweets' variable is truthy and, if so, assigns the authors to the 'authors' variable. If 'parentTweets' is not truthy, the code skips this step and continues with the rest of the function.