"Populate author field in retweet response and update tweetHandlers & apiOperations for debugging"
This commit includes changes to the `tweets.js` controller, `tweetHandlers.js`, and `apiOperations.js` files. In `tweets.js`, the `updatedTweet` object is populated with the `author` field before being returned in the response. In `tweetHandlers.js`, the `retweetCount` variable is no longer being logged to the console. Additionally, the `id` variable is being logged to the console in `apiOperations.js`, but this has been commented out in a later version.