"Refactor: Replace `currentUser` with `user` in tweet bookmarking"
This commit refactors the code related to bookmarking a tweet in the `tweets.js` controller. It replaces all instances of `currentUser` with `user` in the function `bookmarkTweet`. The refactoring is also applied to the helper function `findTweetAndCurrentUser` in `helperFunc.js`, where it now returns `user` instead of `currentUser`. The purpose of this refactoring is to maintain consistency and improve readability.