Change Database.connect() to a static property
The `Database.connect()` method has been changed to a static property named `Database`. The `module.exports` now exports the `Database` class instead of calling the `connect()` method. This allows the `Database` class to be imported and used in other modules, while still maintaining the ability to establish a connection to the MongoDB database using the `connect` static property.