"Update token in likeTweet, retweet, bookmarkTweet, and deleteTweet API responses with a call to tweetManager.saveUser()"
This diff updates the token in the responses of the likeTweet, retweet, bookmarkTweet, and deleteTweet APIs by calling tweetManager.saveUser() instead of generating a new JWT token with updated user information and setting it in the session. This change affects the token returned in the response body, but does not affect the functionality of the APIs.