"Add delete animation for removed tweets and update CSS transitions"
This commit includes changes to both the CSS and JavaScript files. In the CSS file, a new keyframes animation called "deleteAnimation" is added. This animation changes the opacity and vertical position of an element over 0.5 seconds. A new class called "delete-animation" is also added, which applies this animation to any element it is added to.

In the JavaScript file, a setTimeout function is added to delay the removal of a tweet from the DOM after it has been deleted. This allows the delete animation to play before the tweet is removed. The tweet element is given the "delete-animation" class before it is removed, triggering the animation.