"Adds bg-transparent class and updates ContentSection to use it; updates main.css with bg-transparent class definition"
This commit contains changes to both the CSS and JS files. In the CSS file, a new class called ".bg-transparent" is added with a background-color of transparent. In the JS file, the "ContentSection" component's background is updated to use the new ".bg-transparent" class. Additionally, the main.css file is updated with the new class definition.