"Add editTweet endpoint and related functions; update routes"
This diff adds a new `editTweet` function in the `controller/api/tweets.js` file, which is then exported and used in the `routes/api/tweets.js` file to define a new PUT route for editing a tweet. Additionally, the `apiOperations.js` file has a new `toggleEditTweet` function for making API requests to edit a tweet. The `Action.btn.js` file has an updated div for the edit button with a new id.