"Replace `route.route(...).using(...)` with `route.route(...).mid(...)` in several files"
This commit updates the code in several files to replace the use of `route.route(...).using(...)` with `route.route(...).mid(...)`. This change affects the routing of requests in the application and is likely intended to update the syntax to a more recent version of the `xprz` library or a similar routing library. The `mid` method is used to apply middleware to a specific route, similar to how `using` was used in the original code.